Extractor/requirements.txt
black==22.12.0
click==8.1.3
mypy-extensions==0.4.3
numpy==1.24.1
opencv-python==4.7.0.68
pathspec==0.10.3
platformdirs==2.6.2
pyserial==3.5
PySimpleGUI==4.60.4
tomli==2.0.1
typing_extensions==4.4.0
